+++ lib/captain_widget.dart
@@ -12,41 +12,59 @@ import 'package:flutter/widgets.dart';
 /// return a List<Page> depending on current page stack
 class CaptainConfig {
   CaptainConfig({
-    this.pages,
+    required this.pages,
     this.popPage,
     this.shouldPop,
     this.actions,
   });
-
   List<Page> pages;
-  bool popPageRouterFacing(Route<dynamic> _route, dynamic _result) {
-    return popPage(_route, _result, pages);
+  bool popPageRouterFacing(Route<dynamic> _route, dynamic? _result) {
+    return popPage!(_route, _result, pages) ?? _defaultPopPage(_route, _result);
+  }
+
+  bool _defaultPopPage(Route<dynamic> _route, dynamic? _result) {
+    bool _res = _route.didPop(_result);
+    _res ? pages.removeLast() : _res = false;
+    return _res;
   }
 
   @protected
-  final bool Function(Route<dynamic>, dynamic, List<Page>) popPage;
+  final bool? Function(Route<dynamic>, dynamic, List<Page>)? popPage;
 
   Future<bool> shouldPopRouterFacing() async {
-    return await shouldPop(pages);
+    return await shouldPop!(pages) ?? await _defaultShouldPop();
+  }
+
+  Future<bool> _defaultShouldPop() {
+    return SynchronousFuture(pages.isNotEmpty);
   }
 
   @protected
-  final Future<bool> Function(List<Page>) shouldPop;
+  late Future<bool>? Function(List<Page>)? shouldPop;
 
   @protected
-  final Map<dynamic, List<Page> Function(List<Page>)> actions;
+  final Map<dynamic, List<Page> Function(List<Page>)>? actions;
 
   StreamController<bool> shouldRebuildMessengerPipe =
       StreamController.broadcast();
 
   void invokeAction(dynamic actionKey) {
-    pages = actions[actionKey](pages);
+    pages = actions![actionKey]!(pages);
     shouldRebuildMessengerPipe.add(true);
   }
 
   void invokeActionFunc(List<Page> Function(List<Page>) actionFunc) {
     pages = actionFunc(pages);
     shouldRebuildMessengerPipe.add(true);
+  }
+
+  void upsertAction(
+      List<Page> Function(List<Page>) actionFunc, dynamic actionKey) {
+    if (actions!.containsKey(actionKey)) {
+      actions!.update(actionKey, (_) => actionFunc);
+      return;
+    }
+    actions![actionKey] = actionFunc;
   }
 }
 
@@ -57,12 +75,20 @@ extension GetCaptain on NavigatorState {
   }
 
   /// invoke a custom actionFunc on Captain with providing a function that takes a List<Page> as input and returns a List<Page> which will be the new page stack
-  void actionFunc(List<Page> Function(List<Page>) actionFunc) {
+  void actionFunc(
+    List<Page> Function(List<Page>) actionFunc, {
+    dynamic? registerWithFollowingKey,
+  }) {
+    if (registerWithFollowingKey != null) {
+      Captain.of(this.context)
+          .config
+          .upsertAction(actionFunc, registerWithFollowingKey);
+    }
     return Captain.of(this.context).config.invokeActionFunc(actionFunc);
   }
 }
 
-/// Captain widget "Router"
+/// ### Captain widget "Router"
 ///
 /// place the Captain widget beneath your MaterialApp and pass it a CaptainConfig
 /// - Captain widgets can be nested just like Router
@@ -73,11 +99,11 @@ extension GetCaptain on NavigatorState {
 /// ````
 class Captain extends InheritedWidget {
   Captain({
-    @required this.config,
+    required this.config,
   }) : super(
           child: Router(
-            routerDelegate: AppRouterDelegate(
-              config: config,
+            routerDelegate: _AppRouterDelegate(
+              config,
             ),
           ),
         );
@@ -90,15 +116,15 @@ class Captain extends InheritedWidget {
   }
 
   static Captain of(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<Captain>();
+    return context.dependOnInheritedWidgetOfExactType<Captain>()!;
   }
 }
 
-class AppRouterDelegate extends RouterDelegate
+class _AppRouterDelegate extends RouterDelegate
     with ChangeNotifier, PopNavigatorRouterDelegateMixin {
-  AppRouterDelegate({
+  _AppRouterDelegate(
     this.config,
-  }) {
+  ) {
     config.shouldRebuildMessengerPipe.stream.listen((_) {
       notifyListeners();
     });