+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2-null-safety] - 2/12/2021
 
-* TODO: Describe initial release.
+* introduced null-safety + made popPage, shouldPop, actions optional/nullable
+
+## [0.0.2] - 2/12/2021
+
+* updated README.md with code examples + added upserting function for actionFunc extension method
+
+## [0.0.1] - 2/11/2021
+
+* initial release containing Captain Widget + Config Class